Started adding points functionality - Now need to ask users to bet and then do all the betting calculations
