+++ src/main/java/com/pluralsight/UserInterfaces/PointsScreen.java
@@ -1,11 +1,11 @@
 package com.pluralsight.UserInterfaces;
 
 import com.pluralsight.Models.Player;
+import com.pluralsight.Utilities.Utilities;
 
 import java.util.List;
 
 public class PointsScreen extends Screen{
-    private int points;
 
     public PointsScreen(List<Player> players) {
         super(players);
@@ -14,17 +14,20 @@ public class PointsScreen extends Screen{
     //Methods
     @Override
     public void display() {
-        System.out.print("How many points would you like to wager?\n\nEnter amount here: ");
-        int wageredPoints = scanner.nextInt();
+        for (Player player : players) {
+            System.out.println(Utilities.centerMessage(player.getName(), 46, ' '));
+            Utilities.createLineofChars(46, '=');
+            System.out.print("How many points would you like to buy in with?\n\nEnter amount here: ");
+            int wageredPoints = scanner.nextInt();
+            processStartingPoints(player, wageredPoints);
+        }
+    }
+
+
+    //Methods
+    private void processStartingPoints(Player player, int wageredPoints) {
+        player.setPoints(wageredPoints);
     }
 
     //Getters and Setters
-
-    public int getPoints() {
-        return points;
-    }
-
-    public void setPoints(int points) {
-        this.points = points;
-    }
 }
